fix : stdafx.h 포함 및 네임스페이스 오류 해결
