+++ Engine/src/ECS/ComponentSystem.cpp
@@ -1,1 +1,2 @@
+#include "stdafx.h"
 #include "ComponentSystem.h"
