+++ Engine/src/ECS/PhysicsSystem.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "PhysicsSystem.h"
 
 using namespace KGCA41B;
