+++ Engine/src/ECS/System.cpp
@@ -1,1 +1,2 @@
+#include "stdafx.h"
 #include "System.h"
