+++ Engine/src/Engine/DataTypes.h
@@ -1,7 +1,5 @@
 #pragma once
 #include "stdafx.h"
-#include <reactphysics3d/reactphysics3d.h>
-#include <d3d11.h>
 
 namespace KGCA41B
 {
@@ -145,6 +143,30 @@ namespace KGCA41B
 
 		IDLE
 	};
+}
+
+static std::wstring to_mw(const std::string& _src)
+{
+	USES_CONVERSION;
+	return std::wstring(A2W(_src.c_str()));
+}
+
+static std::string to_wm(const std::wstring& _src)
+{
+	USES_CONVERSION;
+	return std::string(W2A(_src.c_str()));
+}
+
+static std::vector<std::string> split(std::string input, char delimiter) {
+	std::vector<std::string> answer;
+	std::stringstream ss(input);
+	std::string temp;
+
+	while (getline(ss, temp, delimiter)) {
+		answer.push_back(temp);
+	}
+
+	return answer;
 }
