+++ Engine/src/Engine/FbxLoader.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "FbxLoader.h"
 
 namespace KGCA41B {
