+++ Engine/src/Engine/Log.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "Log.h"
 
 #include "spdlog/sinks/stdout_color_sinks.h"
