+++ Engine/src/Engine/Scene.cpp
@@ -1,1 +1,2 @@
+#include "stdafx.h"
 #include "Scene.h"
