+++ Engine/src/Engine/SingletonClass/DataMgr.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "common.h"
 #include "DllMacro.h"
 
 namespace KGCA41B
