+++ Engine/src/Engine/SingletonClass/TimeMgr.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "TimeMgr.h"
 
 using namespace KGCA41B;
