+++ Engine/src/Engine/stdafx.h
@@ -1,4 +1,8 @@
 #pragma once
+#include <reactphysics3d/reactphysics3d.h>
+#include "fmod.hpp"
+#include <d3d11.h>
+
 #include <list>
 #include <vector>
 #include <map>
@@ -20,12 +24,10 @@
 #include <functional>
 #include <sstream>
 
-#include "SpacePartition/Collision.h"
-#include "fmod.hpp"
 
-using namespace std;
 using namespace Microsoft::WRL;
 using namespace DirectX;
+using namespace std;
 
 #define SINGLETON(type)\
 public:\
@@ -39,27 +41,3 @@ type() {}\
 ~type() {}\
 type(const type&) = delete;\
 type& operator=(const type&) = delete;
-
-static std::wstring to_mw(const std::string& _src)
-{
-	USES_CONVERSION;
-	return std::wstring(A2W(_src.c_str()));
-}
-
-static std::string to_wm(const std::wstring& _src)
-{
-	USES_CONVERSION;
-	return std::string(W2A(_src.c_str()));
-}
-
-static std::vector<std::string> split(std::string input, char delimiter) {
-	std::vector<std::string> answer;
-	std::stringstream ss(input);
-	std::string temp;
-
-	while (getline(ss, temp, delimiter)) {
-		answer.push_back(temp);
-	}
-
-	return answer;
-}
